feat(json): explicit JSONVal type and helpers
