+++ graphistry/compute/predicates/is_in.py
@@ -1,7 +1,7 @@
 from typing import Any, List
 import pandas as pd
 
-from graphistry.util import assert_json_serializable
+from graphistry.utils.json import assert_json_serializable
 
 from .ASTPredicate import ASTPredicate
